Improve SPADS cluster launcher script for running as systemd service (#7)
The regular behavior of systemd is to when stopping a service send
SIGTERM to all processes in the service cgroup which also kills the
ongoing games. To prevent that we change the kill mode to kill only the
script, that is then properly terminating all SPADS room processes and
waits for the ongoing games to finish.

The launcher script also supports reload: restart the script and SPADS
manager process without touching any of the room instances.